+++ src/testing/helpers/MockOffchainRelayer.sol
@@ -10,11 +10,10 @@ import "forge-std/Vm.sol";
 import "./DeliveryInstructionDecoder.sol";
 import "./ExecutionParameters.sol";
 import "./BytesParsing.sol";
+
 using BytesParsing for bytes;
 
-
 contract MockOffchainRelayer {
-
     uint16 chainIdOfWormholeAndGuardianUtilities;
     IWormhole relayerWormhole;
     WormholeSimulator relayerWormholeSimulator;
@@ -23,9 +22,9 @@ contract MockOffchainRelayer {
 
     mapping(uint16 => address) wormholeRelayerContracts;
 
-    mapping(uint16 => uint) forks;
+    mapping(uint16 => uint256) forks;
 
-    mapping(uint => uint16) chainIdFromFork;
+    mapping(uint256 => uint16) chainIdFromFork;
 
     mapping(bytes32 => bytes[]) pastEncodedVMs;
 
@@ -38,21 +37,15 @@ contract MockOffchainRelayer {
         chainIdOfWormholeAndGuardianUtilities = relayerWormhole.chainId();
     }
 
-    function getPastEncodedVMs(
-        uint16 chainId,
-        uint64 deliveryVAASequence
-    ) public view returns (bytes[] memory) {
+    function getPastEncodedVMs(uint16 chainId, uint64 deliveryVAASequence) public view returns (bytes[] memory) {
         return pastEncodedVMs[keccak256(abi.encodePacked(chainId, deliveryVAASequence))];
     }
 
-    function getPastDeliveryVAA(
-        uint16 chainId,
-        uint64 deliveryVAASequence
-    ) public view returns (bytes memory) {
+    function getPastDeliveryVAA(uint16 chainId, uint64 deliveryVAASequence) public view returns (bytes memory) {
         return pastEncodedDeliveryVAA[keccak256(abi.encodePacked(chainId, deliveryVAASequence))];
     }
 
-    function registerChain(uint16 chainId, address wormholeRelayerContractAddress, uint fork) public {
+    function registerChain(uint16 chainId, address wormholeRelayerContractAddress, uint256 fork) public {
         wormholeRelayerContracts[chainId] = wormholeRelayerContractAddress;
         forks[chainId] = fork;
         chainIdFromFork[fork] = chainId;
@@ -66,18 +59,13 @@ contract MockOffchainRelayer {
         relay(logs, bytes(""));
     }
 
-    function vaaKeyMatchesVAA(
-        VaaKey memory vaaKey,
-        bytes memory signedVaa
-    ) internal view returns (bool) {
+    function vaaKeyMatchesVAA(VaaKey memory vaaKey, bytes memory signedVaa) internal view returns (bool) {
         IWormhole.VM memory parsedVaa = relayerWormhole.parseVM(signedVaa);
-        return (vaaKey.chainId == parsedVaa.emitterChainId)
-            && (vaaKey.emitterAddress == parsedVaa.emitterAddress)
+        return (vaaKey.chainId == parsedVaa.emitterChainId) && (vaaKey.emitterAddress == parsedVaa.emitterAddress)
             && (vaaKey.sequence == parsedVaa.sequence);
     }
 
     function relay(Vm.Log[] memory logs, bytes memory deliveryOverrides) public {
-       
         uint16 chainId = chainIdFromFork[vm.activeFork()];
         require(wormholeRelayerContracts[chainId] != address(0), "Chain not registered with MockOffchainRelayer");
 
@@ -86,9 +74,7 @@ contract MockOffchainRelayer {
         Vm.Log[] memory entries = relayerWormholeSimulator.fetchWormholeMessageFromLog(logs);
         bytes[] memory encodedVMs = new bytes[](entries.length);
         for (uint256 i = 0; i < encodedVMs.length; i++) {
-            encodedVMs[i] = relayerWormholeSimulator.fetchSignedMessageFromLogs(
-                entries[i], chainId
-            );
+            encodedVMs[i] = relayerWormholeSimulator.fetchSignedMessageFromLogs(entries[i], chainId);
         }
         IWormhole.VM[] memory parsed = new IWormhole.VM[](encodedVMs.length);
         for (uint16 i = 0; i < encodedVMs.length; i++) {
@@ -129,9 +115,8 @@ contract MockOffchainRelayer {
     ) internal {
         (uint8 payloadId,) = parsedDeliveryVAA.payload.asUint8Unchecked(0);
         if (payloadId == 1) {
-            DeliveryInstruction memory instruction =
-                decodeDeliveryInstruction(parsedDeliveryVAA.payload);
-            
+            DeliveryInstruction memory instruction = decodeDeliveryInstruction(parsedDeliveryVAA.payload);
+
             bytes[] memory encodedVMsToBeDelivered = new bytes[](instruction.vaaKeys.length);
 
             for (uint8 i = 0; i < instruction.vaaKeys.length; i++) {
@@ -142,38 +127,30 @@ contract MockOffchainRelayer {
                     }
                 }
             }
-            
-            EvmExecutionInfoV1 memory executionInfo =
-                decodeEvmExecutionInfoV1(instruction.encodedExecutionInfo);
-               
+
+            EvmExecutionInfoV1 memory executionInfo = decodeEvmExecutionInfoV1(instruction.encodedExecutionInfo);
+
             uint256 budget = executionInfo.gasLimit * executionInfo.targetChainRefundPerGasUnused
-                + instruction.requestedReceiverValue  + instruction.extraReceiverValue;
+                + instruction.requestedReceiverValue + instruction.extraReceiverValue;
 
             uint16 targetChain = instruction.targetChain;
-            
+
             vm.selectFork(forks[targetChain]);
-           
+
             vm.deal(address(this), budget);
 
-            IWormholeRelayerDelivery(wormholeRelayerContracts[targetChain]).deliver{
-                value: budget
-            }(
-                encodedVMsToBeDelivered,
-                encodedDeliveryVAA,
-                payable(address(this)),
-                deliveryOverrides
+            IWormholeRelayerDelivery(wormholeRelayerContracts[targetChain]).deliver{value: budget}(
+                encodedVMsToBeDelivered, encodedDeliveryVAA, payable(address(this)), deliveryOverrides
             );
-            
+
             setInfo(
                 parsedDeliveryVAA.emitterChainId,
                 parsedDeliveryVAA.sequence,
                 encodedVMsToBeDelivered,
                 encodedDeliveryVAA
             );
-            
         } else if (payloadId == 2) {
-            RedeliveryInstruction memory instruction =
-                 decodeRedeliveryInstruction(parsedDeliveryVAA.payload);
+            RedeliveryInstruction memory instruction = decodeRedeliveryInstruction(parsedDeliveryVAA.payload);
 
             DeliveryOverride memory deliveryOverride = DeliveryOverride({
                 newExecutionInfo: instruction.newEncodedExecutionInfo,
@@ -181,34 +158,24 @@ contract MockOffchainRelayer {
                 redeliveryHash: parsedDeliveryVAA.hash
             });
 
-            EvmExecutionInfoV1 memory executionInfo =
-                decodeEvmExecutionInfoV1(instruction.newEncodedExecutionInfo);
+            EvmExecutionInfoV1 memory executionInfo = decodeEvmExecutionInfoV1(instruction.newEncodedExecutionInfo);
             uint256 budget = executionInfo.gasLimit * executionInfo.targetChainRefundPerGasUnused
                 + instruction.newRequestedReceiverValue;
 
-            bytes memory oldEncodedDeliveryVAA = getPastDeliveryVAA(
-                instruction.deliveryVaaKey.chainId, instruction.deliveryVaaKey.sequence
-            );
-            bytes[] memory oldEncodedVMs = getPastEncodedVMs(
-                instruction.deliveryVaaKey.chainId, instruction.deliveryVaaKey.sequence
-            );
+            bytes memory oldEncodedDeliveryVAA =
+                getPastDeliveryVAA(instruction.deliveryVaaKey.chainId, instruction.deliveryVaaKey.sequence);
+            bytes[] memory oldEncodedVMs =
+                getPastEncodedVMs(instruction.deliveryVaaKey.chainId, instruction.deliveryVaaKey.sequence);
 
-            uint16 targetChain = decodeDeliveryInstruction(
-                relayerWormhole.parseVM(oldEncodedDeliveryVAA).payload
-            ).targetChain;
+            uint16 targetChain =
+                decodeDeliveryInstruction(relayerWormhole.parseVM(oldEncodedDeliveryVAA).payload).targetChain;
 
             vm.selectFork(forks[targetChain]);
-            IWormholeRelayerDelivery(wormholeRelayerContracts[targetChain]).deliver{
-                value: budget
-            }(
-                oldEncodedVMs,
-                oldEncodedDeliveryVAA,
-                payable(address(this)),
-                encode(deliveryOverride)
+            IWormholeRelayerDelivery(wormholeRelayerContracts[targetChain]).deliver{value: budget}(
+                oldEncodedVMs, oldEncodedDeliveryVAA, payable(address(this)), encode(deliveryOverride)
             );
         }
     }
 
     receive() external payable {}
 }
-
